requirements.txt
pyeitaa

pyrogram
tgcrypto

aiosqlite
sqlalchemy

uvloop; sys_platform != "win32"

python-dotenv